requirements.txt
alembic==1.0.10
astroid==2.2.5
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0
Flask-Cors==3.0.3
Flask-JWT-Extended==3.18.2
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.1
Flask-Uploads==0.2.1
gunicorn==19.7.1
idna==2.8
isort==4.3.17
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
lazy-object-proxy==1.3.1
lxml==4.3.4
Mako==1.0.10
MarkupSafe==1.1.1
mccabe==0.6.1
pydotenv==0.0.7
PyJWT==1.7.1
pylint==2.3.1
PyMySQL==0.9.3
pyrsistent==0.15.2
python-dateutil==2.8.0
python-docx==0.8.10
python-dotenv==0.9.1
python-editor==1.0.4
pytz==2019.1
redis==2.10.6
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.3
typed-ast==1.3.4
urllib3==1.25.2
weixin-python==0.5.5
Werkzeug==0.14.1
wrapt==1.11.1
